backend/requirements.txt
fastapi==0.117.1
uvicorn==0.37.0
python-multipart==0.0.20
pydantic==2.11.9
python-dotenv==1.1.1
requests==2.32.5
httpx==0.28.1
